Optionally handle `inner_stats_` being empty
To make it easier to optionally drop `inner_stats_` from the first step
of the dictionary learning Dask graph, go ahead and assign `None` to the
keys and ensure `inner_stats_` is an empty `tuple`. This way
`inner_stats_A` and `inner_stats_B` are generated by scikit-learn
internally and no additional complexity is added to the Dask graph.