Ingest list of TIFF files
While we are able to handle regex in filenames to make it easy to
specify multiple files, in some cases it is useful to specify the exact
ordering of image files we want to load. Here we adjust our TIFF file
handling to allow either filenames with regex or lists of filenames.