Drop `nanshe_workflow.proj`
`nanshe_workflow.proj` contains computations that were designed for
ipyparallel and later retrofitted to work with distributed. However
these are no longer really needed as they have been rewritten to run
using dask directly. So go ahead and remove this module. This should be
also be a useful smoke test to see where else it is still being
imported.